+++ package.json
@@ -21,38 +21,38 @@
   },
   "license": "ISC",
   "dependencies": {
-    "antd": "^3.10.1",
-    "drizzle": "^1.3.3",
-    "drizzle-react": "^1.2.0",
-    "drizzle-react-components": "^1.2.1",
-    "prop-types": "^15.6.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "truffle-hdwallet-provider": "^1.0.2"
+    "antd": "^3.25.3",
+    "drizzle": "^1.4.0",
+    "drizzle-react": "^1.3.0",
+    "drizzle-react-components": "^1.4.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "truffle-hdwallet-provider": "^1.0.17"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/polyfill": "^7.2.5",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.4",
-    "babel-plugin-import": "^1.10.0",
+    "@babel/core": "^7.7.4",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/preset-env": "^7.7.4",
+    "@babel/preset-react": "^7.7.4",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-import": "^1.13.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "bignumber.js": "^8.0.1",
-    "css-loader": "^2.1.0",
+    "bignumber.js": "^9.0.0",
+    "css-loader": "^3.2.0",
     "html-webpack-plugin": "^3.2.0",
-    "less": "^3.9.0",
-    "less-loader": "^4.1.0",
-    "react-hot-loader": "^4.6.3",
-    "sol-merger": "^0.1.3",
-    "style-loader": "^0.23.1",
-    "truffle": "^5.0.0-beta.2",
-    "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.9",
-    "webpack-merge": "^4.2.1"
+    "less": "^3.10.3",
+    "less-loader": "^5.0.0",
+    "react-hot-loader": "^4.12.18",
+    "sol-merger": "^1.1.1",
+    "style-loader": "^1.0.0",
+    "truffle": "^5.1.1",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
+    "webpack-merge": "^4.2.2"
   }
 }
